revert changes to risc0/circuit/recursion/build.rs

+++ risc0/circuit/recursion/build.rs
@@ -52,25 +52,15 @@ fn download_zkr() {
         return;
     }
 
-    // Check the ZKR archive in the out dir and the src dir. Use the out file if it matches the
-    // hash above, but also warn the developer if there exists a copy in the src dir and it does
-    // not match.
-    let zkr_out_valid = out_path.exists() && check_sha2(&out_path);
-
-    if src_path.exists() {
-        if check_sha2(&src_path) {
-            if !zkr_out_valid {
-                fs::copy(&src_path, &out_path).unwrap();
-            }
+    if out_path.exists() {
+        if check_sha2(&out_path) {
             return;
-        } else {
-            println!(
-                "cargo::warning={SRC_PATH} exists in src dir, but does not match the expected hash"
-            );
         }
+        fs::remove_file(&out_path).unwrap();
     }
 
-    if zkr_out_valid {
+    if src_path.exists() && check_sha2(&src_path) {
+        fs::copy(&src_path, &out_path).unwrap();
         return;
     }
